FDG-7680 add site error boundary
