+++ package.json
@@ -122,6 +122,7 @@
     "react-chartjs-2": "^5.2.0",
     "react-device-detect": "^1.17.0",
     "react-dom": "^17.0.2",
+    "react-error-boundary": "^4.0.11",
     "react-fns": "^1.4.0",
     "react-google-recaptcha": "^2.1.0",
     "react-helmet": "^6.1.0",
