+++ src/components/pageError/page-error-graphic.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { pxToNumber } from '../../helpers/styles-helper/styles-helper';
 import { breakpointLg } from '../../variables.module.scss';
-import * as styles from './notFound.module.scss';
+import * as styles from './page-error.module.scss';
 
 const NotFoundGraphic = () => {
     const png = '/not-found/404gif2.png';
