+++ src/components/pageError/page-error.module.scss
@@ -48,6 +48,10 @@ $lead-text-max: 550px;
     width: 100vw;
   }
 
+  .fallbackGraphic {
+    margin: 1rem;
+  }
+
   @media screen and (max-width: $breakpoint-sm - 1px) {
     .graphicPlacement {
       transform: translateX(-1rem);
