+++ src/components/siteLayout/siteLayout.jsx
@@ -1,6 +1,8 @@
 import React from "react";
 import SiteHeader from "../site-header/site-header";
 import SiteFooter from "../site-footer/site-footer";
+import { ErrorBoundary } from "react-error-boundary";
+import PageErrorText from '../../components/pageError/page-error-text';
 
 export const preProdEnvMsg =
   'Loading metadata and data from endpoints in pre-production environment.';
@@ -14,9 +16,11 @@ const SiteLayout = ({ children, isPreProd, glossaryEvent, glossaryClickEventHand
 
   return (
     <div>
-        <SiteHeader lowerEnvMsg={lowerEnvMsg} glossaryEvent={glossaryEvent} glossaryClickEventHandler={glossaryClickEventHandler} />
+      <SiteHeader lowerEnvMsg={lowerEnvMsg} glossaryEvent={glossaryEvent} glossaryClickEventHandler={glossaryClickEventHandler} />
+        <ErrorBoundary FallbackComponent={() => <PageErrorText fallback='true' />}>
           {children}
-        <SiteFooter />
+        </ErrorBoundary>
+      <SiteFooter />
     </div>
   )
 };
